Splitting text in to lines with line prefix
We split the text given on each new line.

Map is creating a new array, for returning an array with the elements
changed, and puts naturally outputs each element on a new line.

Using with_index provides us with the option to give an offset, we use
that with some string interpolation to prefis our lines with "line ##".